refactor: move render-chain composition into core Component
Establish render composition as part of the Component contract rather than
adapter logic. Component's constructor walks the prototype chain once (cached
per class) and installs an own `render` that composes each level leaf->outer,
threading every level's output into its super as `children`. Each level is
called with the receiver as `this`, so an adapter's reactive proxy applies
uniformly across the whole chain via a single call.

This reverts the react adapter to its original single-render form - it now
consumes `this.render` as one composed function with zero chain knowledge,
proving the composition belongs in core. Direct `foo.render()` calls (no
adapter) compose too.

Instance-field `render = fn` initializes after construction and shadows the
composed render; single-level stays identity, multi-level field render is
treated as unsupported (documented inline).

Full mvc and react suites stay green, including the render-chain tests.

Refs #109
